Added login logout pages,also image uploading space

+++ Lang-table.php
@@ -16,24 +16,34 @@
 
     //Showing the Language list
     echo '<h1>Language Table</h1>';
-    echo '<table><thead><th>Language</th><th>Native Speakers</th><th>Country</th><th>linguistic age</th><th>Actions</th></thead>';
-
-    // looping through the data result from the query, and displaying each show name
+    echo '<table><thead><th>Language</th><th>Photo</th><th>Native Speakers</th><th>Country</th><th>linguisticage</th>';
+    if (!empty($_SESSION['username'])) {
+       echo '<th>Actions</th>';
+    }
+     echo '</thead>';
+    // looping through the data result from the query, and displaying each Language name
     foreach ($Language as $Languages) {
         echo '<tr>
         <td>' . $Languages['Language'] . '</td>
+        <td>';
+        if ($Languages['photo'] != null) {
+            echo '<img src="img/uploads/' . $Languages['photo'] . '" />';
+        }
+        echo '</td>
         <td>' . $Languages['NativeSpeakers'] . '</td>
         <td>' . $Languages['Country'] . '</td>
-        <td>' . $Languages['linguisticage'] . '</td>
-        <td>
-        <a href="edit-language.php?languageId=' . $Languages['languageId'] . '">
-            Edit
-        </a>&nbsp;
-        <a href="delete-language.php?languageId=' .$Languages['languageId'] . '" onclick="return confirmDelete();">
-            Delete
-        </a>
-        </td>
-        </tr>';
+        <td>' . $Languages['linguisticage'] . '</td>';
+        if (!empty($_SESSION['username'])) {
+            echo '<td class="actions">
+                <a href="edit-language.php?languageId=' . $Languages['languageId'] . '">
+                    Edit
+                </a>&nbsp;
+                <a href="delete-language.php?languageId=' .$Languages['languageId'] . '" onclick="return confirmDelete();">
+                    Delete
+                </a>
+            </td>';
+        }
+        echo '</tr>';
           }
 
     // end of the list
